requirements.txt
-i https://pypi.org/simple
pillow==9.1.1
python-barcode==0.14.0
